+++ script-a-master/Altv-Roleplay/AltV.Roleplay/models/Server_Items.cs
@@ -23,12 +23,12 @@ namespace Altv_Roleplay.models
         public bool isItemGiveable { get; set; }
         public bool isItemClothes { get; set; }
         public string ClothesType { get; set; }
-        public int ClothesDraw { get; set; }
-        public int ClothesTexture { get; set; }
-        public int ClothesUndershirt { get; set; }
-        public int ClothesUndershirtTexture { get; set; }
-        public int ClothesDecals { get; set; }
-        public int ClothesDecalsTexture { get; set; }
+        public ushort ClothesDraw { get; set; }
+        public byte ClothesTexture { get; set; }
+        public ushort ClothesUndershirt { get; set; }
+        public byte ClothesUndershirtTexture { get; set; }
+        public ushort ClothesDecals { get; set; }
+        public byte ClothesDecalsTexture { get; set; }
         public string itemPicSRC { get; set; }
     }
 }
